Do some cleanup on `PairRequestController#create`
The pattern here will likely repeat itself across multiple
controllers, which doesn't feel great as this is a solid chunk of
code. It might be best to let this play out in a few other places
to make a good abstraction easier to write (maybe a concern?). I
just did a little minimal cleanup