Flash message form errors (#116)
* Add styles for input fields with errors

* Register new flash type for form errors

* Add component for rendering flash messages

* Connect flash component to views and render via turbo streams

* Remove old form error code and fix datetime input's border style

* Extract partial for pair request table row

* Do some cleanup on `PairRequestController#create`

The pattern here will likely repeat itself across multiple
controllers, which doesn't feel great as this is a solid chunk of
code. It might be best to let this play out in a few other places
to make a good abstraction easier to write (maybe a concern?). I
just did a little minimal cleanup

* Add alert classes to Tailwind's safelist

* Fix issue with hidden action column on mobile

* Polish flash container styles