112 edit standup meeting (#120)
* update seeds to create a default standup meeting group

* initial standup_meeting#edit page and form

* shorten standup meeting groups to standups, make it easier to read

* enable skipping a standup meeting

* add completions controller

* remove the stylesheets from the manifest.js

When we use esbuild, it automatically sends the stylesheets to the
builds folder. However, to ensure that flatpickr's styles get picked
up, we import them in the application.js

This can cause a potential sprockets doublelink error. Therefore,
I removed the stylesheets from the manifest.js, which means we will
need to explicitly require the css of additional js dependencies if
they have them

* update specs, use pundit inference pattern

* rubocop fixes

* extract permissions handling to viewcomponent render method

* only make a single db call, use includes to pull association

* restrict accessing the edit form unless you're authorized;
gracefully handle unauthorized error

* group name and date should not be editable on the standup edit form

* remove final reference to @date_range

* update to handle unauthorization in application controller

* use includes instead a second call to get associated relation

---------

Co-authored-by: toyhammered <daniel@rassiner.com>