Profile show page (#207)
* Add basic pundit policy for profiles

* Build layout and styling for profiles show page

Includes a basic edit interface. This isn't how editing will be
realized in the final project. This is just a stand-in to make
profiles easier to work with for the time being.

* Add support for viewing and editing `work_model_preferences`

* Add ability to set `job_search_status`

* Changed the wording for `job_search_status` for clarity

There's not really enough room to communicate that these are job
search statuses on the layout. I think it's good to have job/work
in there somewhere so someone who's just viewing the profile will
catch the meaning.

* Set up profile factory so that picture can be attached

* Add placeholder profile picture

* Add component for rendering the profile picture

* Guard against rendering empty location section

This is because the icon is rendered alone on the page in this
event.

* Bring resume and bragsheet buttons closer together on wide screens

* Pluralize `value` in `work_model_preferences`

Expected type is `Array` here. Makes more sense for it to be
plural

* Call authorization methods for profile editing

* Eager load profile picture attachment and user

* Render `:edit` on failure to update profile